refactor: ACI-5241 group the interactive terminal flows into one package
cmd/common had accumulated fifteen files for the wizard, the browser login, the
form chrome and the prompts, interleaved with the rest of the command layer. They
move to cmd/common/interactive and lose the prefixes that were standing in for a
package name.

hydrateStoredAuth stays behind in cmd/common: the root command calls it and it
needs nothing interactive, which is what keeps the dependency one-way.
