+++ cmd/common/interactive/callback_paste.go
@@ -1,4 +1,4 @@
-package common
+package interactive
 
 import (
 	"bufio"
