+++ cmd/common/interactive/callback_paste_test.go
@@ -1,6 +1,6 @@
 //go:build unit
 
-package common
+package interactive
 
 import (
 	"context"
