+++ cmd/common/interactive/daemon.go
@@ -1,4 +1,4 @@
-package common
+package interactive
 
 import (
 	"context"
