+++ cmd/common/interactive/fix_auth.go
@@ -1,4 +1,4 @@
-package common
+package interactive
 
 import (
 	"context"
