+++ cmd/common/interactive/logout_scope_test.go
@@ -1,6 +1,6 @@
 //go:build unit
 
-package common
+package interactive
 
 import (
 	"testing"
