+++ cmd/common/interactive/wizard_auth.go
@@ -1,4 +1,4 @@
-package common
+package interactive
 
 import (
 	"bufio"
