+++ cmd/common/interactive/wizard_bazel_test.go
@@ -1,6 +1,6 @@
 //go:build unit
 
-package common
+package interactive
 
 import (
 	"os"
