+++ main.go
@@ -6,6 +6,7 @@ import (
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/browse"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/ccache"
 	"github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/common"
+	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/common/interactive" // registers activate --interactive
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/daemon"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/doctor"
 	_ "github.com/bitrise-io/bitrise-build-cache-cli/v3/cmd/file"
